# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.checkCodeRequest:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
  auth.meResponse:
    properties:
      email:
        type: string
      id:
        type: string
      isCreatedProfile:
        type: boolean
      onboardingStep:
        type: string
    type: object
  auth.refreshRequest:
    properties:
      refreshToken:
        type: string
    type: object
  auth.sendCodeRequest:
    properties:
      email:
        type: string
    type: object
  auth.sendCodeResponse:
    properties:
      expiresIn:
        description: seconds
        type: integer
      resendAfter:
        description: seconds
        type: integer
    type: object
  auth.tokenResponse:
    properties:
      accessToken:
        type: string
      isCreatedProfile:
        type: boolean
      onboardingStep:
        type: string
      refreshToken:
        type: string
    type: object
  geo.updateRequest:
    properties:
      lat:
        type: number
      lng:
        type: number
    type: object
  interest.Interest:
    properties:
      id:
        type: string
      value:
        type: string
    type: object
  meeting.Tag:
    properties:
      id:
        type: string
      value:
        type: string
    type: object
  meeting.acceptResponse:
    properties:
      chatId:
        type: string
    type: object
  meeting.adRequest:
    properties:
      description:
        type: string
      tagIds:
        items:
          type: string
        type: array
    type: object
  meeting.adResponse:
    properties:
      authorProfileId:
        type: string
      createdAt:
        type: string
      description:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/meeting.Tag'
        type: array
    type: object
  meeting.feedAuthor:
    properties:
      age:
        type: integer
      avatarUrl:
        type: string
      gender:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  meeting.feedItem:
    properties:
      author:
        $ref: '#/definitions/meeting.feedAuthor'
      description:
        type: string
      distanceKm:
        type: integer
      id:
        description: ad id
        type: string
      tags:
        items:
          $ref: '#/definitions/meeting.Tag'
        type: array
    type: object
  meeting.feedResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/meeting.feedItem'
        type: array
      nextCursor:
        type: string
    type: object
  meeting.likeProfile:
    properties:
      avatarUrl:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  meeting.likeResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      profile:
        $ref: '#/definitions/meeting.likeProfile'
    type: object
  profile.aboutRequest:
    properties:
      datingGoal:
        type: string
      description:
        type: string
      height:
        description: cm, optional
        type: number
      interestIds:
        items:
          type: string
        type: array
      showZodiac:
        type: boolean
      weight:
        description: kg, optional
        type: number
    type: object
  profile.basicsRequest:
    properties:
      birthDate:
        description: YYYY-MM-DD
        type: string
      city:
        type: string
      gender:
        type: string
      name:
        type: string
    type: object
  profile.completeResponse:
    properties:
      isCreatedProfile:
        type: boolean
      onboardingStep:
        type: string
    type: object
  profile.crop:
    properties:
      size:
        type: number
      x:
        type: number
      "y":
        type: number
    type: object
  profile.photoResponse:
    properties:
      crop:
        $ref: '#/definitions/profile.crop'
      id:
        type: string
      position:
        type: integer
      url:
        type: string
    type: object
  profile.profileEnvelope:
    properties:
      onboardingStep:
        type: string
      profile:
        $ref: '#/definitions/profile.profileResponse'
    type: object
  profile.profileResponse:
    properties:
      birthDate:
        type: string
      city:
        type: string
      datingGoal:
        type: string
      description:
        type: string
      gender:
        type: string
      height:
        type: number
      id:
        type: string
      interests:
        items:
          $ref: '#/definitions/interest.Interest'
        type: array
      name:
        type: string
      photos:
        items:
          $ref: '#/definitions/profile.photoResponse'
        type: array
      showZodiac:
        type: boolean
      userId:
        type: string
      weight:
        type: number
    type: object
  profile.reorderRequest:
    properties:
      order:
        items:
          type: string
        type: array
    type: object
  profile.setCropRequest:
    properties:
      crop:
        $ref: '#/definitions/profile.crop'
    type: object
info:
  contact: {}
  description: Backend API for the MeetGo dating app.
  title: MeetGo API
  version: 0.1.0
paths:
  /auth/check_code:
    post:
      consumes:
      - application/json
      parameters:
      - description: email + code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.checkCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.tokenResponse'
        "401":
          description: Unauthorized
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      summary: Verify OTP code and issue tokens
      tags:
      - auth
  /auth/logout:
    post:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Revoke the current session
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: refreshToken
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.tokenResponse'
        "401":
          description: Unauthorized
          schema:
            type: object
      summary: Rotate token pair
      tags:
      - auth
  /auth/send_code:
    post:
      consumes:
      - application/json
      parameters:
      - description: email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.sendCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.sendCodeResponse'
        "422":
          description: Unprocessable Entity
          schema:
            type: object
        "429":
          description: Too Many Requests
          schema:
            type: object
      summary: Send OTP code to email
      tags:
      - auth
  /interests:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/interest.Interest'
            type: array
      security:
      - BearerAuth: []
      summary: List interests reference
      tags:
      - interests
  /likes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/meeting.likeResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Incoming likes on my active ad
      tags:
      - meetings
  /likes/{id}/accept:
    post:
      parameters:
      - description: like id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/meeting.acceptResponse'
        "404":
          description: Not Found
          schema:
            type: object
        "409":
          description: Conflict
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Accept a like (opens a chat)
      tags:
      - meetings
  /likes/{id}/reject:
    post:
      parameters:
      - description: like id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: object
        "409":
          description: Conflict
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Reject a like
      tags:
      - meetings
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.meResponse'
        "401":
          description: Unauthorized
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Current account identity
      tags:
      - auth
  /me/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.profileResponse'
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - profile
  /me/profile/about:
    put:
      consumes:
      - application/json
      parameters:
      - description: about
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/profile.aboutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.profileEnvelope'
        "409":
          description: Conflict
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Onboarding step 2 — about
      tags:
      - profile
  /me/profile/basics:
    put:
      consumes:
      - application/json
      parameters:
      - description: basics
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/profile.basicsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.profileEnvelope'
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Onboarding step 1 — basics
      tags:
      - profile
  /me/profile/complete:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.completeResponse'
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Finalize onboarding (>=2 photos)
      tags:
      - profile
  /me/profile/photos:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: image (jpeg/png/webp, <=10MB)
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/profile.photoResponse'
        "409":
          description: Conflict
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Upload a profile photo (multipart field "photo")
      tags:
      - profile
  /me/profile/photos/{id}:
    delete:
      parameters:
      - description: photo id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete a photo
      tags:
      - profile
    patch:
      consumes:
      - application/json
      parameters:
      - description: photo id
        in: path
        name: id
        required: true
        type: string
      - description: crop
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/profile.setCropRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.photoResponse'
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Set/update a photo crop
      tags:
      - profile
  /me/profile/photos/order:
    patch:
      consumes:
      - application/json
      parameters:
      - description: order
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/profile.reorderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/profile.photoResponse'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Reorder photos (first = main)
      tags:
      - profile
  /meeting-ad:
    post:
      consumes:
      - application/json
      parameters:
      - description: ad
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/meeting.adRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/meeting.adResponse'
        "403":
          description: Forbidden
          schema:
            type: object
        "409":
          description: Conflict
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create my meeting ad
      tags:
      - meetings
  /meeting-ad/{id}:
    get:
      parameters:
      - description: ad id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/meeting.adResponse'
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get an ad by id
      tags:
      - meetings
  /meeting-ad/{id}/like:
    post:
      parameters:
      - description: ad id
        in: path
        name: id
        required: true
        type: string
      responses:
        "201":
          description: Created
        "404":
          description: Not Found
          schema:
            type: object
        "409":
          description: Conflict
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Like an ad
      tags:
      - meetings
  /meeting-ad/me:
    delete:
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove my active ad
      tags:
      - meetings
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/meeting.adResponse'
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get my active ad
      tags:
      - meetings
    put:
      consumes:
      - application/json
      parameters:
      - description: ad
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/meeting.adRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/meeting.adResponse'
        "404":
          description: Not Found
          schema:
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Edit my active ad (description + tags)
      tags:
      - meetings
  /meeting-list:
    get:
      parameters:
      - description: distance | date
        in: query
        name: sort
        type: string
      - description: km
        in: query
        name: radius
        type: number
      - description: MALE | FEMALE
        in: query
        name: gender
        type: string
      - description: 18..99
        in: query
        name: ageMin
        type: integer
      - description: 18..99
        in: query
        name: ageMax
        type: integer
      - description: dating goal
        in: query
        name: goal
        type: string
      - collectionFormat: csv
        description: meeting tag ids
        in: query
        items:
          type: string
        name: tagIds
        type: array
      - description: opaque cursor
        in: query
        name: cursor
        type: string
      - description: page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/meeting.feedResponse'
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Discovery feed of nearby ads (keyset cursor)
      tags:
      - meetings
  /meeting-tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/meeting.Tag'
            type: array
      security:
      - BearerAuth: []
      summary: List meeting tags reference
      tags:
      - meetings
  /my-geo:
    post:
      consumes:
      - application/json
      parameters:
      - description: lat/lng
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/geo.updateRequest'
      responses:
        "204":
          description: No Content
        "422":
          description: Unprocessable Entity
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update my current location (write-only; never returned)
      tags:
      - geo
  /profiles/{id}:
    get:
      parameters:
      - description: profile id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.profileResponse'
        "404":
          description: Not Found
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get a public profile by id
      tags:
      - profile
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"